rename footer
